requirements.txt
bibtexparser>=2.0.0b7 # I need the pre version
flask
Flask-Session
python-box[all]
typer
